Reconocimiento Facial/requirements.txt
numpy==1.26.4
opencv-contrib-python==4.10.0.84
picamera==1.13
picamera2==0.3.12
piexif==1.1.3
piglow==1.2.5
pigpio==1.78
Pillow==8.1.2
RPi.GPIO==0.7.0
RTIMULib==7.2.1

